Update Base overlay with custom class and tweaked TCL
The base overlay needs some reassignment to make
the IOPs and GPIO accessible to end users. This commit
also changes the TCL file to make the microblaze hierarchies
match what is expected by the MicroblazeHierarchy. It also
updates the constants in the pmod and arduino libraries to
match the new TCL