pl.py: hierarchy fixes
1) In .tcl parsing, the gpio_dict local variable is constructed
without fully hierarchical names.  However, when it is read (in
_add_gpio_pins) it is read with fully hierarchical names from
self.nets.  This fixes the construction to match the use.

2) when hierarchies are constructed, they were constructed for every
level of hierarchy which contains IP connected to the PS.  However, if
there is a level of hierarchy without IP connected to the PS, the code
crashed.  The fix ensures that all parent hierarchies are built
regardless of whether they contain IP connected to the PS.