pynqmicroblaze: Perform unsigned conversion correctly and handle hierarchies (#736)
* build_xsdk.tcl:

Allow microblazes to connect to more than one BRAM
Assume that the first BRAM is associated with stdin/stdout

* pl.py: Don't wait forever to disable AXI interfaces

We could wait forever here if there is nothing behind the MMIO object.

* pl.py: hierarchy fixes

1) In .tcl parsing, the gpio_dict local variable is constructed
without fully hierarchical names.  However, when it is read (in
_add_gpio_pins) it is read with fully hierarchical names from
self.nets.  This fixes the construction to match the use.

2) when hierarchies are constructed, they were constructed for every
level of hierarchy which contains IP connected to the PS.  However, if
there is a level of hierarchy without IP connected to the PS, the code
crashed.  The fix ensures that all parent hierarchies are built
regardless of whether they contain IP connected to the PS.

* pynqmicroblaze: fixes for when microblaze hierarchy is not at the toplevel

1) BSP lookup needs to be aware of hierarchical names  (This matches the code in build_sdk.tcl). An alternative would be to lookup with non-hierarchical names, but this would require this code to be fixed and build_sdk.tcl to be fixed.
2) Construction of mb_info needs to be aware of hierarchical names.

* pyprintf: Perform unsigned conversion correctly

The code requires conversting from a 32-bit C number to a python integer.
This conversion has to be aware of the signedness of the C number to perform
correct conversion.  In particular, the new code now matches the C behavior
for cases like:
pyprintf("%x", 0xFFFFFFF0);
The previous code interpreted this as "-16" and output a hex string like "-10"
The new code matches the C output of "FFFFFFF0"