Interrupt Improvements (#1134)
* Add support for directly connected interrupts

Allow interrupt instances be directly connected the PS.

Fixes: Xilinx/PYNQ-Dev#274

* Allow DefaultIP to instantiate in the presence of unconnected Interrupts

At present DefaultIP will through an exception during construction if an interrupt is not connected properly. This change instead raises a warning and sets the interrupt attribute to None. Our current drivers use None as a signal not to use an interrupt so no additional changes are needed.