Move the delay timer to the highest possible number
The delay timer and PWM devices are both currently specified as device 0
meaning that any use of delay will interrupt the PWM the output. By
moving the delay timer to the upper range that will increase the number
of PWM devices that can be used before conflict occurs.

In future we may want to consider a dedicated delay timer in hardware.