Replace xlnk with XRT for embedded allocation (#1235)
* Add cacheable suppot to XRT allocation

* Refactor XRT downloading out of XrtDevice.download

* Allow subclassing of XrtDevice

* Add GROUP_* constants to Xclbin formats

* Add support for group topologies

* Add support for Xclbin and HWH parsers to take in-memory data rather than files

* Move memories in HWH files to the mem dict

This requires changing classes that interact with BRAM to use mem_dict
rather than ip_dict

* Replace Xlnk with XRT

XlnkDevice is removed and EmbeddedDevice is provided instead

* Add cache and pointer support to XRT allocaters

This is required to get decent performance with the video subsystem
which rapidly cycles frames. A related change is updating the VDMA
to support non-default memories. The main reason is that access
.default_memory requires a PL server lookup so has to be done
outside of the critical loop when making use of the cache

* Patch xclbinutil to use the correct format string

On 32-bit architectures xclbinutil segfaults on completion due to a
mismatched format string

* Add XRT setup to bootpy startup script

* Ensure that memories coming from HWH files are always used