Refactor XSAParser
Added Xsa class to handle reding/extracting from XSA archive
Added XsaParser(Xsa) class to contain parsing code

Changes to be committed:
	modified:   pynq/xsaparser.py

creating a utility to take an xsa containing BDCs, along with the project directory, and construct a new XSA with the missing register map information

Grabbing the XCI files for all the IP in the BDCs from the project directory.

Added some classes for capturing the hierarchy of the regmaps, however I'm struggling to parse the XML of the xci files

XCI fully parsed into an abstract internal representation

slight refactoring

grabbing the docstrings for the registers and fields, this frankensteins monster is starting to come together

Has the ability to render the internal metadata representation as a json file that can be appended and packaged up with the XSA file.

Now packages up a bew XSA with the extra metadata

Woops, was not closing the file before archiving, this is why the JSON format was a bit mangled

Getting the path for the bitstream from Geoff's XSA parser.

created a temp fix for the requirement that the base hwh filename needs to match the bitstream. I think in the future we need to undo this requirement, but that is quite a substantial change so need to think carefully about it.